+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Bug fixes:
 * `Kernel#sprintf` encoding validity has been fixed (#1852, @XrXr).
 * Fixed File.fnmatch causes ArrayIndexOutOfBoundsException (#1845).
 * Make `String#concat` work with no or multiple arguments (#1519).
+* Implement Numeric#clone method to return self.
 
 Compatibility: